requirements.txt
fiftyone
google-genai
python-dotenv
matplotlib
pandas